Estrutura: ajustada novo teste e pattern de pastas para o projeto
Com base em novos estudos foi ajustada a estrutura de pastas para se
adequar a uma nova resolução
