+++ src/modules/apiSettings/popups/authorization/functions/index.ts
@@ -1,0 +1,11 @@
+import { processData } from '../../../services/popup/authorization';
+
+export const useAuthPopup = (closeModal: any) => {
+  const handleSaveData = async (values : any) => {
+      console.log(values);
+      await processData(values);
+      closeModal();
+  };
+    
+  return { handleSaveData };
+};
