+++ src/modules/apiSettings/popups/integration/domain/index.ts
@@ -1,0 +1,7 @@
+import { DataType } from '../../../domain';
+
+export type ItgPopupType = {
+    closeModal: () => void;
+    currentItem: any;
+    data: DataType[];
+};
