--- src/modules/apiSettings/popups/itgPopup/domain/index.ts
@@ -1,4 +1,0 @@
-export type AuthPopupType = {
-    closeModal: () => void;
-    currentItem: any;
-};
