--- src/modules/apiSettings/popups/itgPopup/functions/index.ts
@@ -1,10 +1,0 @@
-import { processData } from '../../../services/popup/authPopup';
-
-export const useAuthPopup = (closeModal: any) => {
-  const handleSaveData = async (values : any) => {
-      await processData(values);
-      closeModal();
-  };
-    
-  return { handleSaveData };
-};
