--- src/modules/apiSettings/services/pages/apiSettings/index.ts
@@ -1,10 +1,0 @@
-import { checkMicroservice as checkMicroserviceInfra } from '../../../infra/pages/apiSettings';
-
-/**
- * Checks the microservice by calling the checkMicroserviceInfra function.
- *
- * @return {Promise<boolean>} A Promise that resolves to a boolean indicating whether the microservice check is successful.
- */
-export const checkMicroservice = async (): Promise<boolean> => {
-    return checkMicroserviceInfra();
-}
