+++ src/modules/apiSettings/services/popup/authorization/index.ts
