+++ src/modules/apiSettings/services/popup/integration/index.ts
@@ -1,0 +1,14 @@
+import {
+    insertData, updateData,
+} from '../../../infra/popup/itgPopup';
+
+
+export const processData = async (data: any) => {   
+    console.log(data);
+
+    if (typeof data.ID !== "number") {
+        insertData(data);
+    } else {
+        updateData(data);
+    }
+}
