--- src/modules/apiSettings/services/popup/itgPopup/index.ts
@@ -1,9 +1,0 @@
-//import { insertData, updateData } from "../../infra/popup/requets";
-
-export const processData = async (data: any) => {   
-    /*if (data.ID !== undefined) {
-        insertData(data);
-    } else {
-        updateData(data);
-    }*/
-}
