--- src/modules/dashboard/components/tabs/index.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import ChartGrid from '../../../../common/components/base-layouts/ChartGrid';
-import { chartData } from './constants';
-const Dashboard: React.FC = () => {
-  return (
-      <ChartGrid data={chartData} pageTitle="Dashboard"/>
-  );
-};
-
-export default Dashboard;
