+++ src/modules/dashboard/pages/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Dashboard: React.FC = () => {
+  return (
+      <div />
+  );
+};
+
+export default Dashboard;
