+++ src/modules/routes/index.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 // Import components from modules
-import Dashboard from '../dashboard/components/tabs';
-import ApiConfigPages from '../apiSettings/pages/apiSettings';
+import Dashboard from '../dashboard/pages';
+import ApiConfigPages from '../apiSettings/pages/general';
 import ScreenLayout from '../../common/components/layout/screenLayout';
 import '../i18n/i18n';
-import { menuItems } from '../menu/domain';
+import { menuItems } from '../menu/constants';
 
 function AppRouter() {
   return (
